+++ include/shield/shield.hpp
@@ -3,7 +3,7 @@
 // Usage: #include <shield/shield.hpp>
 
 #include "shield/AhoCorasickScanner.hpp"
-#include "shield/Re2PatternMatcher.hpp"
+#include "shield/PatternMatcher.hpp"
 #include "shield/RuleSet.hpp"
 #include "shield/RuleManager.hpp"
 #include "shield/MaskingEngine.hpp"